exported to common ButtonGoHome and AlertEmailDelay
